add webconfig file and fix minor bugs (#2)

+++ estoque-api/src/main/java/br/com/nexdom/estoque_api/services/ProductService.java
@@ -1,6 +1,7 @@
 package br.com.nexdom.estoque_api.services;
 
 import br.com.nexdom.estoque_api.dto.request.ProductRequest;
+import br.com.nexdom.estoque_api.dto.request.ProductUpdateRequest;
 import br.com.nexdom.estoque_api.dto.response.ProductByTypeResponse;
 import br.com.nexdom.estoque_api.dto.response.ProductProfitResponse;
 import br.com.nexdom.estoque_api.dto.response.ProductResponse;
@@ -8,22 +9,21 @@ import br.com.nexdom.estoque_api.entities.Product;
 import br.com.nexdom.estoque_api.enums.ProductType;
 import br.com.nexdom.estoque_api.mappers.ProductMapper;
 import br.com.nexdom.estoque_api.repositories.ProductRepository;
+import br.com.nexdom.estoque_api.repositories.StockMovementRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ProductService {
     private final ProductRepository productRepository;
+    private final StockMovementRepository stockMovementRepository;
     private final ProductMapper productMapper;
 
     public ProductResponse create(ProductRequest productRequest) {
@@ -46,16 +46,22 @@ public class ProductService {
         return productMapper.toDTOList(products);
     }
 
-    public ProductResponse update(Long id, ProductRequest productRequest) {
-        Product product = productRepository.findById(id)
+    public ProductResponse update(ProductUpdateRequest productUpdateRequest) {
+        Product product = productRepository.findById(productUpdateRequest.id())
                 .orElseThrow(() -> new EntityNotFoundException("Produto não encontrado"));
 
-        productRepository.save(product);
+        Product newProduct = productRepository.save(productMapper.toEntity(productUpdateRequest));
 
-        return productMapper.toDTO(product);
+        return productMapper.toDTO(newProduct);
     }
 
     public void delete(Long id) {
+        boolean hasMovements = stockMovementRepository.existsByProductId(id);
+
+        if (hasMovements) {
+            throw new IllegalStateException("Não é possível excluir um produto que já possui movimentações no estoque.");
+        }
+
         productRepository.deleteById(id);
     }
